pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "todash"
version = "0.1.0"
description = "Vim-like TUI for Todoist"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "rich>=14.3.3",
    "textual>=8.1.1",
]

[project.scripts]
todo = "todash.cli:main"
